FIX: Use Cloudfront in front of API Gateway
ap-southeast-3 doesn't yet have support for using web ACL in front of
API gateway.

Put the web ACL in front of a cloudfront distribution instead, direct
the cloudfront distribution to use the API gateway as the origin, and
disable caching. Change the website's API requests to point at the
cloudfront distribution. Note that the API gateway endpoint is still
available to be queried, but the domain is not made publicly visible.