FIX: Specify region when making s3 cli requests
When upload.sh calls aws s3 sync, it sometimes fails with a client error
(IllegalLocationConstraintException). "The (un)specified location
constraint is incompatible for the region specific endpoint this request
was sent to".

Specify the location of the bucket when calling the aws s3 sync command,
to prevent the script from just using whatever your REGION envionment
variable might be set to at the time.