FIX: Make bcftools access ap-southeast-3 buckets
The new region ap-southeast-3 does not yet have correct redirection
handling set up for s3, so buckets there cannot be accessed through
s3.amazonaws.com, only s3.ap-southeast-3.amazonaws.com. Htslib, that
handles the file access for bcftools, relies on this redirection
capability and so cannot access buckets in the ap-southeast-3 region.

Override the host value for htslib, to force it to use the endpoint for
the region in which the API module is created, which will in this case
always be the correct region as having all the infrastructure in
Indonesia is a requirement. Most beacons will be serving data from
buckets within their region as well, as this reduces costs and latency.
However, it should be noted that bcftools will be unable to access files
in buckets that are in a different region. If we want to implement a fix
to allow this, we will need to record the region of each bucket and use
that to seed the HTS_S3_HOST environment variable before calling bcftools.