Copy environment.ts to environment.prod.ts
Both environment.ts and environment.prod.ts are identical files, but
they are created separately.

While the contents are identical, we might as well only create it once
and copy the file. If we need to differentiate later we can reverse this
change, and we might even look at removing one of the files if we only
use, for example, environment.prod.ts.