FEAT: Create trigger for updateData
The updateData lambda function relies on the trigger to be created
manually, to prevent being called on every uploaded file.

Create the trigger in terraform, but ensure the /submit endpoint is only
called when a particular vcf file is created. The pipeline creates files
in the form gisaid/gisaid01101.vcf.gz, with a binary representation of
the file number in the name. Thus there will always be one file created
in the form gisaid/gisaid0*\.vcf\.gz, so any other file will
short-circuit the lambda function.