FEAT: Allow backend to handle multiple organisms
Backend has the contig names and maximum nucleotides hardcoded to
SARS-CoV-2, and carries residual fuzzy chromosome matching logic from
the original sBeacon implementation.

Allow the assemblies and contigs to be set using terraform variables.
Note that the contig sizes don't need to be exact; they're only used to
ensure that the backend doesn't spin off many performQuery instances in
regions that don't exist. This also removes vcf-specific chromosomes,
which were not being used, and shouldn't be encouraged anyway. If we
need support for multiple VCFs per sample, we'll need to update this
section.
Frontend support for multiple organisms or different contigs is yet to
be developed.