FEAT: Query different assemblies and contigs
Adding additional organisms into the beacon means we have to start
separating ourselves from SARS-CoV-2. Some of these organisms may have
multiple different contigs making up their reference genome, so we need
to represent that in the query.

Add contig to mutation information, separated by the "-" character. The
default is "1", to maintain compatibility with the SARS-CoV-2 dataset.
Also add the assembly, by default "hCoV-19", at the front of the query
string separated by the "~" character. This in particular is a quick and
dirty solution, and should probably be replaced by a dropdown menu at
the start of the query. TODO: Combine the query handling into a single
function, rather than duplicating the argument parsing across two to
four locations.